backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.20
pydantic==2.10.6
google-generativeai==0.8.4
